Change variable name scopes in preparation for deep models.
To be backwards compatible with old TensorFlow models, the version
number has been bumped to 0.2 (and is now written to the config file
during training, like in the Theano version).  If an older model is
used, then the variables are mapped into the new naming scheme when
the model is restored.