Automatically set vocab sizes if not given
Like in the Theano version, if the source or target vocabulary size is
not given on the command-line (for training), then the value is
determined automatically from the corresponding dictionary file.

It is now also possible to set source vocabulary sizes independently for
each factor.  Note that the corresponding option name has been changed
from --source_vocab_size to --source_vocab_sizes.