Add --valid_script option
This allows the user to specify a script for external scoring
of the validation set. Note that usage differs from the
Theano version's --external_validation_script and instead is
similar to Marian's --valid-script-path option, i.e. the
script is no longer responsible for translation, only scoring.
It is given translations of the validation set (in a file
specified by the first argument) and must write a score to
standard output.