Backward-compatibility for converted Theano models
The Theano and TensorFlow versions of Nematus use different bias
terms in their GRUs. This change causes Nematus to switch to the Theano
style when loading a model that was converted from a Theano model
(which is automatically determined from the config file).

Fixes issue #82 (https://github.com/EdinburghNLP/nematus/issues/82)