Transformer support (main credit: Denis Emelin)
This commit adds experimental support for Transformer models. The code
is adapted from Denis Emelin's Nematode project:

    https://github.com/demelin/nematode

Important: it is likely that variable naming etc. will change in the
near future meaning that models built using this code may not be usable
by future versions of Nematus. You may wish to wait until Transformer
support is more stable (when it will be added to master).

Note that some Nematus features are currently not supported by the
Transformer model. This includes factors and mixture of softmaxes,
among others.