Be more consistent about text file encodings
The current approach is to leave the encoding unspecified when opening
files and therefore to rely on locale.getpreferredencoding() for
decoding to text and encoding to output files. The exception is JSON
files, which are always read and written as UTF-8 (the JSON RFC
requires one of UTF-8, UTF-16, or UTF-32, with UTF-8 being recommended).

In practice, the user's locale will probably specify UTF-8. If there is
a mismatch between the user's locale settings and the actual encodings
used for input files then this may result in UnicodeDecodeErrors. Also
if characters in the vocabulary (originally read into Nematus from a
UTF-8 JSON file) cannot be represented in the target encoding, then this
will result in a UnicodeEncodeError.

An alternative would be to read and write UTF-8 everywhere, regardless
of the user's locale settings. If their files were in a different
encoding, they would have to convert them to avoid errors.