Scale gradients according to actual batch size
Actual batch sizes are sometimes smaller than the configured batch size.
For instance, there will usually be a short minibatch at the end of a
maxibatch if we specify a token-based batch size. To prevent these
short minibatches from having a disproportionate effect, we scale the
gradients by the ratio of the actual / configured batch size.