simplify dropout configuration in RNN:
--rnn_use_dropout: removed; always assumed to be true
--rnn_dropout_embedding: default changed from 0.2 to 0.0
--rnn_dropout_hidden: default changed from 0.2 to 0.0

if you use --rnn_use_dropout, you can remove it, but should
double-check whether you're affected by the changes in default values.

you will only see a silent change in behavior if you train a model,
explicitly set --rnn_dropout_* to a value > 0, but do not set
--rnn_use_dropout. Previously, dropout would still be off,
but now, it will be on.